Add files via upload
One-dimensional model of a quantum wave incident on a finite potential which relays the transmission and reflection coefficients. Also note that it relays the probabilities of locating the quantum particle within the potential - and though physically this will always clear, the particle has low energy within the potential reducing the wave velocity to nearly zero. 

Another feature is the mapping to the potential space (or k-space). This indicates both the magnitude (energy) and direction of the wave packet (packets upon transmission and reflection).